make AB_div_mode a double

+++ src/main/coordinates/nmesh_coordinates.h
@@ -274,7 +274,7 @@ int sphere_around_full_box_at_xc(tMesh *mesh, int N,
                                  double xc[3], double dc[3], double r0);
 int sphere_nAB_around_empty_box_at_xc(tMesh *mesh, int N,
                                       double xc[3], double dc[3], double r0,
-                                      int AB_div_mode, int nAB_x[3]);
+                                      double AB_div_mode, int nAB_x[3]);
 int two_spheres_around_box_at_xc(tMesh *mesh, int N,
                                  double xc[3], double dc[3],
                                  double r0, double r1, int stretch);
